added interest

+++ CheckingAccount.java
@@ -2,29 +2,29 @@
 public class CheckingAccount
 {
     // instance variables
-    private int balance;
+    private double balance;
     // constructor
     public CheckingAccount()
     {
         balance = 10;
     }
-    public CheckingAccount(int startingBalance)
+    public CheckingAccount(double startingBalance)
     {
         balance = startingBalance;
     }
     // methods
     // deposit
-    public void deposit(int value)
+    public void deposit(double value)
     {
         balance += value;
     }
     // withdraw
-    public void withdraw(int value)
+    public void withdraw(double value)
     {
         balance -= value;
     }
     // get balance
-    public int getBalance()
+    public double getBalance()
     {
         return balance;
     }
